Revert "Update power profile to match stock values"
This should not have been merged.  Samsung is NOTORIOUS for screwing up power_profile - we've seen tuna power_profile values on Exynos devices as an example.

Among the obvious problems with the commit that is being reverted:
Missing 1.4 and 1.6 GHz entries on a device that supports operation at those frequencies.
Missing cpu.awake value to handle wakelock power calculation (this was added a long time ago by one of the CM maintainers, prior to the file being forked and copied to multiple repos).
Power consumption values for frequencies between 200 and 1200 MHz that match the profile entries for Exynos 4210 EXACTLY, despite the fact that the Exynos 4210 is built on a 45nm process and the 4412 was built using a 32nm process (consumes less power).

This reverts commit d76e08813ff4cefbc7b00b7c077d3b29804f1cd2

Change-Id: I4b42df70593ff4a7175c0e5acb022da2934c7b7e